LLMs/milvus_chatbot/requirements.txt
langchain==0.0.274
text-generation==0.6.0
pymilvus==2.3.0
python-dotenv==1.0.0
grpcio==1.56.0